Leptos 0.5 (#7)
Migrates to the newest release of Leptos.
[0.5.0-beta](https://github.com/leptos-rs/leptos/releases/tag/v0.5.0-beta).

I started by applying the pattern replacements recommended by the
release notes.

From there I had two big hurdles. 

(1) I had to fix how the async actions were called, so that they could
still access the client web socket handle.
(2) I had to fix how the client state is threaded through the whole
program.

Issues 1 & 2, are symptoms of the same thing, they worked fine when the
context scopes were explicit. With them implicit, they run just slightly
afoul of the implicit context relationships. For the (1) the solution
was to use an explicit Owner. For (2) the solution was to refactor so
that the implicit relationships became correct.